+++ opencga-catalog/src/main/java/org/opencb/opencga/catalog/db/mongodb2/CatalogMongoIndividualDBAdaptor.java
@@ -404,7 +404,7 @@ public class CatalogMongoIndividualDBAdaptor extends AbstractCatalogMongoDBAdapt
     public QueryResult<Individual> update(Query query, ObjectMap parameters) { return null; }
 
     @Override
-    public QueryResult<Integer> delete(Query query) {
+    public QueryResult<Long> delete(Query query) {
         return null;
     }
